fix: swap order of checks
it was a bit silly to access an element of an array
*before* checking if the arrary had a positive length...

+++ packages/nitro-protocol/contracts/test/TESTAssetHolder.sol
@@ -50,7 +50,7 @@ contract TESTAssetHolder is AssetHolder {
         external
         override
     {
-        _transfer(channelId, allocationBytes, new uint[](0));
+        _transfer(channelId, allocationBytes, new uint256[](0));
     }
 
     /**
